Merge `main` into `develop`

+++ src/lexer/tokens/number/number.cpp
@@ -1,4 +1,4 @@
 #include "./number.hpp"
 
-NumberToken::NumberToken(const int value)
+NumberToken::NumberToken(const unsigned long long value)
     : Token(TokenType::NUMBER), value(value) {};
